requirements.txt
# Core dependencies
google-generativeai>=0.8.0
google-adk>=1.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
aiofiles>=23.0.0

# Database and storage
sqlite3
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web scraping and search
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
duckduckgo-search>=3.9.0

# Audio processing (TTS/STT)
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
wave>=0.0.2

# File processing
pypdf2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pillow>=10.0.0

# Data analysis (optional)
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Browser automation (optional)
playwright>=1.40.0

# Terminal and system
psutil>=5.9.0
subprocess32>=3.5.4
ptyprocess>=0.7.0

# Web UI
jinja2>=3.1.0
static3>=0.7.0
markdown>=3.5.0

# Utilities
colorama>=0.4.6
rich>=13.6.0
click>=8.1.0
pyyaml>=6.0.1
jsonschema>=4.19.0
httpx>=0.25.0

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
